JWT token authentication is integrated into the main application in the api with books

+++ src/routers/book_router.py
@@ -1,13 +1,18 @@
-from typing import Annotated, List, Optional
-from fastapi import APIRouter, Depends, UploadFile, status
+from typing import List, Optional
+from fastapi import APIRouter, UploadFile, status
 
+from auth.validation import check_book_author
+from constants import ALLOWED_FILE_TYPES, REQUIRED_FILE_FIELDS
 from custom_exceptions.book_exceptions import BookBulkImportException
-from dependencies import FiltersDep, SessionDep, BookRepositoryDep, SortingDep
-from schemas.auth_schemas import UserOut
+from dependencies import (
+    ValidateBatchSizeDep, FiltersDep,
+    SessionDep, BookRepositoryDep, SortingDep, 
+    UserDep, ValidateBookIdDep
+)
 from schemas.book_schemas import BookCreateSchema, BookNewSchema, BookSchema, BookUpdateSchema
 from utils.enums import OnErrorEnum
 from utils.util_funcs import parse_file, split_into_batches
-from utils.validation_funcs import validate_batch_size, validate_book_data, validate_book_id
+from utils.validation_funcs import validate_book_data
 
 
 router = APIRouter(
@@ -36,7 +41,7 @@ async def get_books(
 async def get_book_by_id(
     session: SessionDep,
     book_repo: BookRepositoryDep,
-    book_id: int = Depends(validate_book_id),
+    book_id: ValidateBookIdDep,
 ) -> BookSchema:
     return await book_repo.get_book_by_id(
         session=session,
@@ -48,31 +53,37 @@ async def create_book(
     session: SessionDep,
     book_repo: BookRepositoryDep,
     book_insert: BookCreateSchema,
-    author: Annotated[UserOut, Depends(get_current_auth_user)],
+    author: UserDep,
 ) -> BookNewSchema:
     return await book_repo.create_book(
         session=session,
-        book_insert=book_insert
+        book_insert=book_insert,
+        author_id=author.id
     )
 
-@router.post('/import/', status_code=status.HTTP_201_CREATED)
+@router.post(
+    '/import/', 
+    status_code=status.HTTP_201_CREATED, 
+    description=f'Required fields: {", ".join(REQUIRED_FILE_FIELDS)}. Allowed file types: {", ".join(ALLOWED_FILE_TYPES)}'
+)
 async def bulk_import_books(
     session: SessionDep,
     book_repo: BookRepositoryDep,
     file: UploadFile,
+    author: UserDep,
+    batch_size: ValidateBatchSizeDep,
     on_validation_error: OnErrorEnum = OnErrorEnum.RAISE_ERROR,
-    batch_size: int = Depends(validate_batch_size),
 ):
     try:
         data = await parse_file(file)
 
         for batch in split_into_batches(data, batch_size):
-            validated_data: list[BookCreateSchema] = validate_book_data(batch, on_validation_error)
+            validated_data: list[dict] = validate_book_data(batch, on_validation_error)
 
             if not validated_data:
                 return {"message": "empty"}
             
-            await book_repo.create_books_bulk(session, validated_data)
+            await book_repo.create_books_bulk(session, validated_data, author.id)
 
         return {"message": "success"}
 
@@ -86,20 +97,27 @@ async def update_book(
     session: SessionDep,
     book_repo: BookRepositoryDep,
     book_update: BookUpdateSchema,
-    book_id: int = Depends(validate_book_id),
+    book_id: ValidateBookIdDep,
+    author: UserDep,
 ) -> BookNewSchema:
+    await check_book_author(session, book_repo, book_id, author)
+
     return await book_repo.update_book(
         session=session,
         book_update=book_update,
         book_id=book_id,
+        author_id=author.id
     )
 
 @router.delete('/{book_id}/', status_code=status.HTTP_204_NO_CONTENT)
 async def delete_book(
     session: SessionDep,
     book_repo: BookRepositoryDep,
-    book_id: int = Depends(validate_book_id),
+    book_id: ValidateBookIdDep,
+    author: UserDep,
 ) -> None:
+    await check_book_author(session, book_repo, book_id, author)
+
     return await book_repo.delete_book(
         session=session,
         book_id=book_id
